Added support for [] operator.
This support will allow us to use the JSON format in a much more JavaScript
like fashion. This way we can use that format for many files we deal with
in the Snap! C++ environment instead of XML which is always very difficult
to deal with.

This allows for a complete tree, opposed to our config files (which do
also allow for a complete tree, but with the price of having to build
a complete key from the root).