Added LibUtf8 as a required dependency to compile.
Updated the headers to make use of #pragma once.
Renamed all the classes with lower case and separate words by '_'.
Updated the JSON tests so they work (which includes quite a bit of the lexer).
Changed the stream implementation to allow the use of any standard stream.
Implemented a few cases in the catch_stream.cpp file.
Renamed exceptions without the word "exception_" at the start.
Renamed the exit exception "as2js_exit()" and moved the exit code at the end (like in advgetopt).
Simplified the match_pattern() sub-class with a simple static function.
Changed the as2js::String to a basic C++ string. We only support UTF-8 now (through libutf8).
Renamed the int64 as integer and the float64 as floating_point.
Fixed the is_floating_point() and to_floating_point() functions (string to float) testing all possible error cases.
Also fixed the simplify() function when parsing a floating point.
Fixed a bug with the fact that a backslash needs to be escaped properly and in our case it was not.
Fixed the JSON test to make use of the test g_tmp_dir instead of "/tmp/...".
Fixed the RC get_home() so when the pointer is NULL it doesn't crash.
Added the last include in the entire library.
Applied various clean ups when noticing things that could be cleaned up.
Changed the license to GPL v3.