Added the as-rc tool.
This tool is used to transform resource files to .c or .cpp files so you
can directly compile that data in your project (no need to load a file).

Added the AsRcConfig.cmake so we can use a simple cmake function to convert
resource files.

Use as-rc to generate the license.cpp/.h files linked in the tools.