Generate the list of identifier characters.
I changed the way I maintain the list of characters that can compose an
identifier. JavaScript now uses the Unicode definition nearly to the letter.
That means a very large number of characters can be used in an identifier.

I had a unicode_characters tool in the tests folder. I now use that tool to
generate the list of characters acceptable in an identifier. At some point,
though, I plan to have that in the libutf8. For now, it uses the ICU which
is problematic to have code that works across platforms (i.e. older versions
of the OS have older versions of the ICU and thus older versions of the
Unicode standard... if you want the same script to work across versions,
you should not use Unicode characters. Limit yourself to ACSII and it will
work always.)