Got the NATIVE flag to flow properly.
I think that my prepare_attributes() function is still quite wrong. Many
of the flags need to be going through or not depending on many more details
not yet taken in account properly. But the NATIVE flag is really important
since I need it to know how to compile/optimize the code.