Many more changes to get the compiler to work.
This includes tests for all the unary & binary operators except the ':='
and [] which I will work on next.

Moved the functions used to do the search of an identifier in one file
called resolver instead of scattered by type (class, function, expression,
package).

Renamed the rc class and file as resources.

Renamed the db file as database (like the class).

Fixed some issues in the resources implementation to properly handle the
list of script paths (it's not a string anymore, it's a list of strings.)

Although the new compiler test now passes, the changes broke many old tests
so I'll also have to update that... It's all coming together though.