Finally go the auto-enumeration to work again!
Fixed the prepare_attributes(), we also cannot skip on node_t::NODE_PACKAGE.
Fixed some direct calls to the node->get_attribute(...) which needed to be get_attribute(node, ...).
Added a verification target to make sure the compiler does not use node->get_attribute() which is wrong (except in the prepare_attributes()).
Added a call_operator() to convert native operators back to operators if defined as MEMBER calls.
Fixed the +a optimization because +"" returns NaN and not "".
Verify number of parameters for getters and setters.
Various simplifications & clean ups.