Added support for an archive for run-time funcs.
Got the '**' operator implemented. It compiles properly.
So I now have '+' (binary), '-' (binary), '*', '**' and '~' implemented.
Just some 20 other operators to work on. But first I'll write a test that
can verify the output and for that I want a "run environment" (i.e. be
able to load a compiled expression and execute it on the fly).