Added the double multiplicative.
Removed the need for the runtime library as it stands.
Instead, I'll move the running_file class to that library so one does not have
to run against the entire as2js library (although I have to see what
dependencies are required because it's likely that as2js would be one of those
dependencies).