Got the new strings tests to pass.
Fixed a memcpy() that used ptr + <offset> where ptr was of the wrong type.
Fixed a string flag. It was set to ALLOCATED but the data was set to nullptr.
Removed test of f_data_size when copying strings. This is variable.
Properly copy strings when ALLOCATED is set, and if not set and long, generate an error.
Added support for boolean type in generate_assignment().
Marked variable in meta of additive strings as strings.
Support quoted strings in .meta files.
Added full support for strings in the binary tests.
Removed many debug std::cerr...