Added string compare operators.
Added a relocation to the data size (useful for strings).
Fixed the 1 byte relocations; they are after the 8 byte relocations and the offset needed to be adjusted.
Fixed a MOV which used 0x45 instead of 0x85 (to get a disp32).
Fixed handling of NaN when applying the logical not to a floating point.