Implemented the `length` field.
The `length` is a field so it works 99% like the `[]` operator. It is
now implemented as expected within the generate_array(). It is one MOV
instruction, so super well optmized.

Changed the RANGE support to a separate `Range` type. It is not yet
implemented but it will make a lot more sense that way.

Fixed typing of members so it would work better (and it does!)

Moved the loading of the String.length size to a separate reusable function.

Small clean ups.