Implemented the String.charAt() function.
This version has some ways to call a native function. I already implemented
one for the String class. This includes fixes and cleanups in the compiler
and parser and node. Additions to support the CALL properly. And a test
to run the charAt() function and verify that it works as expected.

The test may not pass just yet, but the binary.sh returns the correct results.