Updated README to make it a bit more accurate.
Restore the change to the string.cpp; the migrate_cpp tool had an issue when
it timed out it would not run the clean command. I think with that fix we're
very close to perfection without the need to have an explicit rm before
building the next item.

Added another test with a digit followed by a letter other than x or X.