Added assignments for strings.
Added a few more native functions for String type. That allowed me to find
a few bugs which I fixed along the way. Especially, empty lists must just
be empty. No way we want to add a LIST object when it's empty.

Started changing the String functions to only use Integers since doubles
are otherwise interpreted in "a weird way" (in some cases, as a subscript,
it fails, and in some cases it works, better force users to make sure they
have integers instead).