Added indexOf() and lastIndexOf() for String.
This version is capable of compiling the indexOf() and lastIndexOf() functions.
I had to change the handling of a list of parameters to a CALL because we need
to save all the results in an ARRAY and each must be a binary_variable. That
meant native types, such as Integer, needed some additional work.

Fixed/cleaned up a few other things along the way.