Added some additional VARIABLE flags.
In order for the CALL to work, it needs results in a binary VARIABLE and to
do so it sets a flag to true. When the flatting happens, that flag is used
to set the VARIABLE flag of the temporary variable to true. This is now done
for many instructions.