Added 25 math functions.
These functions are all in the C library except for fround() which I took
from the Node.js implementation.

The tests pass, however, many are just returning NaN or +/-Inf. We should
look at using valid inputs for all to make sure it works as expected.

I added many nodes, so it is not likely that all the tests pass. I'll do that
next. Then later I'll add the few additional functions (with 2+ parameters).