Implemented clz32() and hypot().
The hypot() is "childish" in the sense that there are no optimizations.
If you use very large numbers, it will fail when there are ways to still
make it work. At the same time, it is written inline so it is really fast.

The clz32() function uses the BSR instruction and "flip" the number (31 - BSR
result; force 32 if input is zero).